shared/pyproject.toml
[tool.poetry]
name = "shared"
version = "0.1.1"
description = "A package containing utilities shared between the dataprocessing and the backend side of the project"
authors = ["ChidiRNweke <chidi125@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pgvector = "^0.2.5"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
